+++ src/Controllers/SpecificController.php
@@ -207,10 +207,14 @@ class SpecificController extends AbstractBaseController implements
             }
 
             foreach ($model->getValues() as $key => $value) {
+                #if (! $value instanceof SpecificValueModel) {
+                #    continue;
+                #}
+
                 $metaValue             = null;
                 $defaultValueAvailable = false;
 
-                $value->getSpecificId()->setEndpoint($model->getId()->getEndpoint());
+                #$value->getSpecificId()->setEndpoint($model->getId()->getEndpoint());
 
                 //Get i18n
                 foreach ($value->getI18ns() as $i18n) {